pipeline/requirements.txt
openai
pinecone-client
python-dotenv
pytube
moviepy
google-api-python-client